Refactoring for linear.
